+++ Source/Core/Duality/Components/Rendering/Lights/SpotLightComponent.cs
@@ -84,7 +84,7 @@ namespace Duality.Graphics.Components
 
 		void UpdateLight()
 		{
-			Light.Position.Set(this.GameObj.Transform.Pos.X, this.GameObj.Transform.Pos.Y, this.GameObj.Transform.Pos.Z);
+			Light.Position.Set((float)this.GameObj.Transform.Pos.X, (float)this.GameObj.Transform.Pos.Y, (float)this.GameObj.Transform.Pos.Z);
 
 			Light.Color = new THREE.Math.Color(Color.R / 255f, Color.G / 255f, Color.B / 255f);
 			Light.Intensity = Intensity;
@@ -94,7 +94,7 @@ namespace Duality.Graphics.Components
 			Light.Angle = Angle;
 
 			Vector3 forward = GameObj.Transform.Forward;
-			Light.Target.Position.Set(this.GameObj.Transform.Pos.X + forward.X, this.GameObj.Transform.Pos.Y + forward.Y, this.GameObj.Transform.Pos.Z + forward.Z);
+			Light.Target.Position.Set((float)this.GameObj.Transform.Pos.X + (float)forward.X, (float)this.GameObj.Transform.Pos.Y + (float)forward.Y, (float)this.GameObj.Transform.Pos.Z + (float)forward.Z);
 
 			Light.CastShadow = CastShadow;
 			Light.Shadow.Camera.Near = NearClip;
@@ -121,7 +121,7 @@ namespace Duality.Graphics.Components
 			Light.Shadow.Camera.Fov = Fov;
 			Light.Shadow.MapSize.Set(512, 512);
 
-			Light.Position.Set(this.GameObj.Transform.Pos.X, this.GameObj.Transform.Pos.Y, this.GameObj.Transform.Pos.Z);
+			Light.Position.Set((float)this.GameObj.Transform.Pos.X, (float)this.GameObj.Transform.Pos.Y, (float)this.GameObj.Transform.Pos.Z);
 
 			Light.Color = new THREE.Math.Color(Color.R / 255f, Color.G / 255f, Color.B / 255f);
 			Light.Intensity = Intensity;
@@ -131,7 +131,7 @@ namespace Duality.Graphics.Components
 			Light.Angle = Angle;
 
 			Vector3 forward = GameObj.Transform.Forward;
-			Light.Target.Position.Set(this.GameObj.Transform.Pos.X + forward.X, this.GameObj.Transform.Pos.Y + forward.Y, this.GameObj.Transform.Pos.Z + forward.Z);
+			Light.Target.Position.Set((float)this.GameObj.Transform.Pos.X + (float)forward.X, (float)this.GameObj.Transform.Pos.Y + (float)forward.Y, (float)this.GameObj.Transform.Pos.Z + (float)forward.Z);
 
 			Scene.ThreeScene.Add(Light);
 			Scene.ThreeScene.Add(Light.Target);
