+++ Source/Core/Duality/Components/Rendering/SkyboxComponent.cs
@@ -78,9 +78,9 @@ namespace Duality.Graphics.Components
 
 		void UpdateSkybox()
 		{
-			sky.Position.Set(this.GameObj.Transform.Pos.X, this.GameObj.Transform.Pos.Y, this.GameObj.Transform.Pos.Z);
-			sky.Rotation.Set(this.GameObj.Transform.Rotation.X, this.GameObj.Transform.Rotation.Y, this.GameObj.Transform.Rotation.Z, THREE.Math.RotationOrder.XYZ);
-			sky.Scale.Set(this.GameObj.Transform.Scale.X, this.GameObj.Transform.Scale.Y, this.GameObj.Transform.Scale.Z);
+			sky.Position.Set((float)this.GameObj.Transform.Pos.X, (float)this.GameObj.Transform.Pos.Y, (float)this.GameObj.Transform.Pos.Z);
+			sky.Rotation.Set((float)this.GameObj.Transform.Rotation.X, (float)this.GameObj.Transform.Rotation.Y, (float)this.GameObj.Transform.Rotation.Z, THREE.Math.RotationOrder.XYZ);
+			sky.Scale.Set((float)this.GameObj.Transform.Scale.X, (float)this.GameObj.Transform.Scale.Y, (float)this.GameObj.Transform.Scale.Z);
 
 			(sky.Material as THREE.Materials.ShaderMaterial).Uniforms["turbidity"] = new GLUniform { { "value", Turbidity } };
 			(sky.Material as THREE.Materials.ShaderMaterial).Uniforms["rayleigh"] = new GLUniform { { "value", Rayleigh } };
